requirements.txt
pip                 24.3.1
python-dotenv       1.2.1
python-telegram-bot 13.15
redis               7.1.0
vk_api              11.10.0
